+++ src/main/kotlin/com/nylas/util/JsonHelper.kt
@@ -30,6 +30,11 @@ class JsonHelper {
       .add(UpdateWhenAdapter())
       .add(CreateWhenAdapter())
       .add(IMessageAdapter())
+      .add(CreateConnectorAdapter())
+      .add(CredentialDataAdapter())
+      .add(MicrosoftAdminConsentCredentialDataAdapter())
+      .add(GoogleServiceAccountCredentialDataAdapter())
+      .add(ConnectorOverrideCredentialDataAdapter())
       // Polymorphic adapters
       .add(WHEN_JSON_FACTORY)
       .add(FREE_BUSY_JSON_FACTORY)